TODO: Relocations need to be in u8 units

+++ crates/lsh/src/compiler/tokenizer.rs
@@ -8,7 +8,7 @@ use std::str::Chars;
 pub enum Token<'a> {
     // Literals
     Identifier(&'a str),
-    Integer(i32),
+    Integer(u32),
     Regex(&'a str),
 
     // Keywords
@@ -195,7 +195,7 @@ impl<'a> Tokenizer<'a> {
         }
 
         let s = &self.input[self.start_pos..self.current_pos];
-        match s.parse::<i32>() {
+        match s.parse::<u32>() {
             Ok(val) => Token::Integer(val),
             Err(_) => Token::Error(format!("Invalid integer: {s}")),
         }
